requirements.txt
django==5.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
channels-redis==4.0.0
flask==2.3.3
daphne==4.0.0
beautifulsoup4==4.10.0
chardet==5.2.0
django_probes==1.7.0